Add 'ffmpeg' feature which is disabled by default To enable this feature pls build with: cargo build --release --features ffmpeg
