+++ opengb/src/directors/global_state.rs
@@ -2,8 +2,9 @@ use super::PersistentState;
 use crate::asset_manager::AssetManager;
 use crate::utilities::StoreExt2;
 use radiance::{
-    audio::{AudioSource, AudioSourceState, Codec},
-    media::{MediaEngine, VideoSource},
+    audio::{AudioEngine, AudioSource, AudioSourceState, Codec as AudioCodec},
+    rendering::VideoPlayer,
+    video::Codec as VideoCodec,
 };
 use regex::Regex;
 use std::{
@@ -20,9 +21,9 @@ pub struct GlobalState {
 
     asset_mgr: Rc<AssetManager>,
     bgm_source: Box<dyn AudioSource>,
-    movie_source: Box<dyn VideoSource>,
     sound_sources: Vec<Rc<RefCell<Box<dyn AudioSource>>>>,
     default_scene_bgm: HashMap<String, String>,
+    video_player: Box<VideoPlayer>,
 
     pass_through_wall: bool,
 }
@@ -30,11 +31,11 @@ pub struct GlobalState {
 impl GlobalState {
     pub fn new(
         asset_mgr: Rc<AssetManager>,
-        media_engine: &Rc<dyn MediaEngine>,
+        audio_engine: &Rc<dyn AudioEngine>,
         persistent_state: Rc<RefCell<PersistentState>>,
     ) -> Self {
-        let bgm_source = media_engine.create_audio_source();
-        let movie_source = media_engine.create_video_source(asset_mgr.component_factory());
+        let bgm_source = audio_engine.create_source();
+        let video_player = asset_mgr.component_factory().create_video_player();
         let sound_sources = vec![];
         let music_path = "/basedata/basedata/datascript/music.txt";
         let default_scene_bgm =
@@ -47,9 +48,9 @@ impl GlobalState {
             role_controlled: 0,
             asset_mgr,
             bgm_source,
-            movie_source,
             sound_sources,
             default_scene_bgm,
+            video_player,
             pass_through_wall: false,
         }
     }
@@ -70,14 +71,9 @@ impl GlobalState {
         self.role_controlled = role_controlled
     }
 
-    pub fn play_movie(&mut self, name: &str) -> (u32, u32) {
-        let data = self.asset_mgr.load_movie_data(name);
-        self.movie_source.play(data, false)
-    }
-
     pub fn play_bgm(&mut self, name: &str) {
         let data = self.asset_mgr.load_music_data(name);
-        self.bgm_source.play(data, Codec::Mp3, true);
+        self.bgm_source.play(data, AudioCodec::Mp3, true);
     }
 
     pub fn play_default_bgm(&mut self) {
@@ -111,8 +107,15 @@ impl GlobalState {
         self.bgm_source.as_mut()
     }
 
-    pub fn movie_source(&mut self) -> &mut dyn VideoSource {
-        self.movie_source.as_mut()
+    pub fn video_player(&mut self) -> &mut VideoPlayer {
+        self.video_player.as_mut()
+    }
+
+    pub fn play_movie(&mut self, name: &str) -> Option<(u32, u32)> {
+        let data = self.asset_mgr.load_movie_data(name);
+        let factory = self.asset_mgr.component_factory();
+        self.video_player
+            .play(factory, data, VideoCodec::Bik, false)
     }
 
     pub fn asset_mgr(&self) -> Rc<AssetManager> {
