+++ opengb/src/directors/sce_vm.rs
@@ -3,9 +3,8 @@ use crate::{asset_manager::AssetManager, loaders::sce_loader::SceFile};
 use encoding::{DecoderTrap, Encoding};
 use imgui::*;
 use log::{debug, error, warn};
-use radiance::input::InputEngine;
-use radiance::media::MediaEngine;
 use radiance::scene::{Director, SceneManager};
+use radiance::{audio::AudioEngine, input::InputEngine};
 use std::fmt::Debug;
 use std::{
     any::Any,
@@ -35,7 +34,7 @@ pub struct SceVm {
 
 impl SceVm {
     pub fn new(
-        media_engine: Rc<dyn MediaEngine>,
+        audio_engine: Rc<dyn AudioEngine>,
         input_engine: Rc<RefCell<dyn InputEngine>>,
         sce: SceFile,
         sce_name: String,
@@ -45,7 +44,7 @@ impl SceVm {
     ) -> Self {
         let state = SceState::new(
             input_engine.clone(),
-            media_engine.clone(),
+            audio_engine.clone(),
             asset_mgr.clone(),
             Rc::new(sce),
             sce_name,
@@ -1018,13 +1017,13 @@ pub struct SceState {
     run_mode: i32,
     ext: HashMap<String, Box<dyn Any>>,
     input_engine: Rc<RefCell<dyn InputEngine>>,
-    media_engine: Rc<dyn MediaEngine>,
+    audio_engine: Rc<dyn AudioEngine>,
 }
 
 impl SceState {
     pub fn new(
         input_engine: Rc<RefCell<dyn InputEngine>>,
-        media_engine: Rc<dyn MediaEngine>,
+        audio_engine: Rc<dyn AudioEngine>,
         asset_mgr: Rc<AssetManager>,
         sce: Rc<SceFile>,
         sce_name: String,
@@ -1040,7 +1039,7 @@ impl SceState {
             run_mode: 1,
             ext,
             input_engine,
-            media_engine,
+            audio_engine,
         }
     }
 
@@ -1060,8 +1059,8 @@ impl SceState {
         self.input_engine.borrow()
     }
 
-    pub fn media_engine(&self) -> &Rc<dyn MediaEngine> {
-        &self.media_engine
+    pub fn audio_engine(&self) -> &Rc<dyn AudioEngine> {
+        &self.audio_engine
     }
 
     pub fn run_mode(&self) -> i32 {
