+++ radiance/src/lib.rs
@@ -9,10 +9,10 @@ pub mod audio;
 pub mod imgui;
 pub mod input;
 pub mod math;
-pub mod media;
 pub mod radiance;
 pub mod rendering;
 pub mod scene;
+pub mod video;
 
 mod constants;
