+++ tools/dev_tools/src/directors/components/audio_pane.rs
@@ -2,10 +2,7 @@ use crate::directors::DevToolsState;
 
 use super::ContentPane;
 use audio::{AudioSourceState, Codec};
-use radiance::{
-    audio::{self, AudioSource},
-    media::MediaEngine,
-};
+use radiance::audio::{self, AudioEngine, AudioSource};
 use std::path::PathBuf;
 
 pub struct AudioPane {
@@ -15,13 +12,13 @@ pub struct AudioPane {
 
 impl AudioPane {
     pub fn new(
-        media_engine: &dyn MediaEngine,
+        audio_engine: &dyn AudioEngine,
         data: Vec<u8>,
         codec: Option<Codec>,
         path: PathBuf,
     ) -> Self {
         let source = codec.map(|c| {
-            let mut source = media_engine.create_audio_source();
+            let mut source = audio_engine.create_source();
             source.play(data, c, true);
             source
         });
